MethaneGraphicsApp library changed from INTERFACE to STATIC. AppCameraController implementation moved to cpp.
